extract sudo calls from script as as_root
On some systems the user I want to run these scripts on are not in the
sudoers group, but might be in the docker group.

Additionally, this is useful to check whether all of these elevated
commands are actually needed.

Using this commit I'm testing this by calling the script setting the
environment variable SUDO_CMD to an empty string. For example:

    SUDO_CMD= bin/build.sh