add bin/appimg.sh to build an app image out of i2p and divachain
This script expects bin/build.sh to have been successfully run and
copies the result from the docker build out of the images into the
build/exchange.diva.divachain.AppDir/
which is the basis for the build using the appimagetool called at the
end of the script.

This is done because the same i2p and divachain build process should be
used for the app image as well. So in this POC we just use the built
binaries of the respecting containers. Because of this we need to resort
to binary patching to adjust these binaries for them to be able to be
run inside of an AppImage instead a docker container environment.

The AppRun script is the entrypoint which gets executed when starting
the resulting appimage.

The idea is that this script setups up the required data directories
required by i2p, starts the i2p daemon & starts divachain.

When closing divachain the i2p daemon should be stopped as well.